Add Avatto thermostats (Tuya) (#3748)
